Add option for restart by kwarg
Save log files to dir unless given with a path in the kwarg in
ftp_download.
Add keyword argument restart defaulting to "ask" to ftp_download.
If restart is set to the answer whether to continue a previous session
("y", "n" or "l"), there is no interruption of the current session if
the download log file has a standard name. Otherwise, the is a prompt
for the download log file name or, if restart is set to "ask", for a
choice of continuation.
Add args dir and restart to init_restart and adjust the function for the
new restart option.
Update documentation.