Refactor download process
Revise download process relying more on the inventory.
Update all routines related to the creation and processing of the
inventory and move them to their own source file `inventory.jl`.
Move all types to a new file `types.jl`. Add a new ConnectionError and a
struct file that saves all relevant path information for a file.
Reintroduce Printf for transforming float version numbers to strings.
Remove the Bits dependency. Remove the init function and the conversion
tool and point to the internal conversion script instead.
Update licence information in the README.

Signed-off-by: Peter Bräuer <pb866.git@gmail.com>