Save new extension in inventory
Save the newext to the metadata of the inventory. Rename the converted
count to conversions.
Refactor the code to include functions from the conversion script rather
than running the script itself to convert files. Include a new script
for functions omitting conversions, if opted out.
Rename data to assets folder.

Signed-off-by: Peter Bräuer <pb866.git@gmail.com>