Integrate h4toh5 functions in ICARE.jl
Revise strategy to run conversion script to include functions directly
and allow their overloading. Add a function newext that returns the
extension of the desired converted output format.
Fix remotefiles! to add the original extension to the inventory, if it
was previously empty. This fixes problems with downloading files from
the server.
Change converter kwarg to a flag convert, which can be switched off, if
conversion is not desired.
Bump SFTP to version 0.2.1.

Signed-off-by: Peter Bräuer <pb866.git@gmail.com>