[#16] Prepare file conversion, add db clean-up
Add option erase to clean function to remove original or converted files
from the database with Extension enum options original, converted, none.
Add combine argument to localscan to allow different set operations,
which will allow scanning for the downloaded part of the database or for
files not belonging to the database. This prepares file conversion.

Signed-off-by: Peter Bräuer <pb866.git@gmail.com>