[#16] Overload Base.convert, update documentation
Instead of defining and exporting a custom convert function, overlaod
Base.convert. Define convert as public rather than exporting it as it is
already available in Base.
Update the documentation about conversion and add a new inventory page.
Add a source for the ICARE package in the docs Project.toml.

Signed-off-by: Peter Bräuer <pb866.git@gmail.com>