[#17] Refine attach! method, fixes
Refine path handling. Also save parent folders of a given path to extras
to ensure the path is not deleted during clean if the parent folders are
not part of the inventory.
Ensure, inventory is saved to file by updating the updated time.
Save only relative paths in the inventory.

Signed-off-by: Peter Bräuer <pb866.git@gmail.com>