[#17] Add function list_inventory, minor improvements
Add funktion list_inventory to show a simplified tree structure of the
inventory together with statistics.
Prefer AbstractLogger over ConsoleLogger in function signatures to be
more flexible. Prefer SortedDict for inventory function arguments
without type specifications.
Fix setting resync to true when the inventory has a temp section at a
point where the inventory is not defined yet.
Add a new entry for file downloads to the metadata of the inventory.

This closes #17.

Signed-off-by: Peter Bräuer <pb866.git@gmail.com>